feat: use context to determine mobile vs desktop datepicker behaviour
