+++ frontend/src/components/DatePicker2/components/DatePickerCalendar.tsx
@@ -1,0 +1,22 @@
+import { Calendar } from '~components/Calendar'
+
+import { useDatePicker } from '../DatePickerContext'
+
+export const DatePickerCalendar = (): JSX.Element => {
+  const {
+    colorScheme,
+    internalValue,
+    isDateUnavailable,
+    handleDateChange,
+    initialFocusRef,
+  } = useDatePicker()
+  return (
+    <Calendar
+      colorScheme={colorScheme}
+      value={internalValue ?? undefined}
+      isDateUnavailable={isDateUnavailable}
+      onChange={handleDateChange}
+      ref={initialFocusRef}
+    />
+  )
+}
