+++ frontend/src/components/DatePicker2/index.ts
@@ -1,3 +1,2 @@
-export type { DatePickerProps } from './DatePicker'
-export { DatePickerContainer as DatePicker } from './DatePickerContainer'
+export type { DatePicker, DatePickerProps } from './DatePicker'
 export type { DatePickerBaseProps } from './types'
